+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.5.0
+
+* Preserve image format in PIL.Image.Image when loading
+* Added ONNX version of Detectron2 and make default model
+* Remove API code, we don't serve this as a standalone API any more
+* Update ordering logic to account for multicolumn documents.
+
 ## 0.4.4
 
 * Fixed patches not being a package.